fix: comprehensive code review fixes (critical + high + medium)
Phase 1 - Critical:
- Fix PUSH_I8/U8/I16/U16 emitting wrong byte counts in Compiler
- Add VM bytecode bounds checking for all instructions
- Fix ASTNodeType enum value conflicts (TAGGED_VALUE_EXPRESSION, PACK_EXPRESSION)
- Fix unterminated string infinite loop in Lexer
- Fix AST memory leaks (PropertyDef destructor, NewTypeDef genericParams)
- Fix null pointer dereference in repr() methods

Phase 2 - High:
- Fix withStream catching all exceptions instead of LexException
- Fix CustomizedType::match() to compare moduleId when both present
- Fix ParserImpl storing ParseState by value (changed to reference)
- Add documentation for Untyped::match() design intent

Phase 3 - Medium:
- Add overflow-checked arithmetic (checked_add/sub/mul/negate)
- Fix insert(begin()) O(n²) pattern in VM
- Fix LexState::next() parameter type (int → size_t)
- Fix tokenType variable shadowing in Lexer
- Replace fragile enum range checks with Set lookups
- Fix CMAKE_BUILD_TYPE to be a cache variable
- Update .clang-format Standard to c++23
- Fix accept() EOF error messages
- Fix revert() misleading default parameter
- Fix typos (moudleName, isTermintator, exponentalSet, etc.)

Tests: 687/687 passing, +16 new test cases

+++ src/parsing/LexState.cpp
@@ -36,7 +36,7 @@ namespace NG::parsing
     return source.at(index + 1);
   }
 
-  void LexState::next(int n)
+  void LexState::next(size_t n)
   {
     if (!eof())
     {
